Add client and payment entities and their repositories into project (#32)
Add client and payment entities. Add client and payment repositories. Update maven dependencies. Update payment units.
