+++ payment-domain/src/main/java/rd/epam/java/payment/domain/entity/Client.java
@@ -1,6 +1,7 @@
 package rd.epam.java.payment.domain.entity;
 
 import lombok.Data;
+import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
@@ -22,14 +23,15 @@ import java.util.UUID;
 @Table(name = "pm_clients")
 public class Client {
     @Id
-    @GeneratedValue
+    @GeneratedValue(generator = "UUID")
+    @GenericGenerator(
+            name = "UUID",
+            strategy = "org.hibernate.id.UUIDGenerator")
     @Column(name = "client_id")
     private UUID id;
 
-    @Column(name = "name")
     private String name;
 
-    @Column(name = "description")
     private String description;
 
     @OneToMany(fetch = FetchType.EAGER, mappedBy = "client", cascade = CascadeType.ALL)
