+++ payment-domain/src/main/java/rd/epam/java/payment/domain/entity/Payment.java
@@ -1,6 +1,7 @@
 package rd.epam.java.payment.domain.entity;
 
 import lombok.Data;
+import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
@@ -21,31 +22,27 @@ import java.util.UUID;
 @Table(name = "pm_payments")
 public class Payment {
     @Id
-    @GeneratedValue
-    @Column(name="payment_private_id")
+    @GeneratedValue(generator = "UUID")
+    @GenericGenerator(
+            name = "UUID",
+            strategy = "org.hibernate.id.UUIDGenerator")
     private UUID paymentPrivateId;
 
-    @Column(name="payment_public_id")
     private UUID paymentPublicId;
 
-    @Column(name="client_id")
-    @ManyToOne(cascade= CascadeType.ALL)
-    @JoinColumn(name="client_id")
+    @ManyToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "client_id")
     private Client clientId;
 
-    @Column(name="receiver_account_id")
-    @OneToOne (cascade=CascadeType.ALL)
-    @JoinColumn(name="receiver_account_id")
+    @OneToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "receiver_account_id")
     private UUID receiverAccountId;
 
-    @Column(name="sender_account_id")
-    @OneToOne (cascade=CascadeType.ALL)
-    @JoinColumn(name="sender_account_id")
+    @OneToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "sender_account_id")
     private UUID senderAccountId;
 
-    @Column(name="amount")
     private Double amount;
 
-    @Column(name="key")
     private String key;
 }
