+++ payment-impl/src/main/java/rd/epam/java/payment/repository/PaymentRepository.java
@@ -1,6 +1,8 @@
+package rd.epam.java.payment.repository;
+
 import lombok.extern.slf4j.Slf4j;
 
-import rd.epam.java.payment.domain.entity.Client;
+import rd.epam.java.payment.domain.entity.Payment;
 
 import javax.persistence.EntityManager;
 import javax.persistence.Persistence;
@@ -10,24 +12,24 @@ import java.util.List;
 import java.util.Optional;
 
 /**
- * Client repository for working with client table in database
+ * Payment repository for working with payment table in database
  *
  * @author Dmitrii_Lopatin
  */
 @Slf4j
-public class ClientRepository {
+public class PaymentRepository {
     private final EntityManager entityManager = Persistence.createEntityManagerFactory("payment-unit").createEntityManager();
 
     /**
-     * Add client record into database
+     * Add payment record into database
      *
-     * @param client - record for saving
+     * @param payment - record for saving
      */
-    public void save(Client client) {
+    public void save(Payment payment) {
         try {
-            log.info("save() - save client {}", client);
+            log.info("save() - save payment {}", payment);
             entityManager.getTransaction().begin();
-            entityManager.persist(client);
+            entityManager.persist(payment);
             entityManager.getTransaction().commit();
         } catch (Exception e) {
             log.warn("Error during saving: ", e);
@@ -35,16 +37,16 @@ public class ClientRepository {
     }
 
     /**
-     * Finds client in database by his id
+     * Finds payment in database by its id
      *
-     * @param id of client
-     * @return found client or empty object
+     * @param id of payment
+     * @return found payment or empty object
      */
-    public Optional<Client> findById(Integer id) {
+    public Optional<Payment> findById(Integer id) {
         try {
-            log.info("findById() - find client by id = {}", id);
-            Client client = entityManager.find(Client.class, id);
-            return Optional.ofNullable(client);
+            log.info("findById() - find payment by id = {}", id);
+            Payment payment = entityManager.find(Payment.class, id);
+            return Optional.ofNullable(payment);
         } catch (Exception e) {
             log.warn("Error during searching by id: ", e);
         }
@@ -52,15 +54,15 @@ public class ClientRepository {
     }
 
     /**
-     * Finds list of clients in database
+     * Finds list of payments in database
      *
      * @param ids - list of id
-     * @return list of clients or empty list
+     * @return list of payments or empty list
      */
-    public List<Client> findByIdList(List<Integer> ids) {
+    public List<Payment> findByIdList(List<Integer> ids) {
         try {
-            log.info("findByIdList() - find list of clients by id = {}", ids);
-            TypedQuery<Client> query = entityManager.createQuery("Select b from pm_clients b Where b.client_id=:ids", Client.class);
+            log.info("findByIdList() - find list of payments by id = {}", ids);
+            TypedQuery<Payment> query = entityManager.createQuery("Select b from pm_payments b Where b.payment_private_id=:ids", Payment.class);
             return query.setParameter("ids", ids).getResultList();
         } catch (Exception e) {
             log.warn("Error during searching by id list: ", e);
@@ -69,18 +71,18 @@ public class ClientRepository {
     }
 
     /**
-     * Updates one client record in database
+     * Updates one payment record in database
      *
-     * @param client for updating
-     * @return updated client or empty object
+     * @param payment for updating
+     * @return updated payment or empty object
      */
-    public Optional<Client> update(Client client) {
+    public Optional<Payment> update(Payment payment) {
         try {
-            log.info("update() - update client {}", client);
+            log.info("update() - update payment {}", payment);
             entityManager.getTransaction().begin();
-            entityManager.merge(client);
+            entityManager.merge(payment);
             entityManager.getTransaction().commit();
-            return Optional.of(client);
+            return Optional.of(payment);
         } catch (Exception e) {
             log.warn("Error during updating: ", e);
         }
@@ -88,15 +90,15 @@ public class ClientRepository {
     }
 
     /**
-     * Deletes one client record
+     * Deletes one payment record
      *
      * @param id of record for deleting
      */
     public void delete(Integer id) {
         try {
-            log.info("delete() - delete client by id = {}", id);
+            log.info("delete() - delete payment by id = {}", id);
             entityManager.getTransaction().begin();
-            entityManager.remove(entityManager.find(Client.class, id));
+            entityManager.remove(entityManager.find(Payment.class, id));
             entityManager.getTransaction().commit();
         } catch (Exception e) {
             log.warn("Error during deleting: ", e);
